Minor refactoring that looks like a huge deal. Its only renaming of mother package and organizing of child pom's.
