+++ deploy-manager-admin/pom.xml
@@ -4,14 +4,14 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
   <parent>
-    <groupId>org.springframework.boot</groupId>
-    <artifactId>spring-boot-starter-parent</artifactId>
-    <version>1.2.5.RELEASE</version>
+    <artifactId>no.difi.deploymanager</artifactId>
+    <groupId>difi-deploy-manager</groupId>
+    <version>0.9.1-SNAPSHOT</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>no.difi.deploymanager.admin</groupId>
-  <artifactId>difi.deploy.manager.admin</artifactId>
+  <artifactId>deploy-manager-admin</artifactId>
   <version>0.9.1-SNAPSHOT</version>
 
   <properties>
