+++ deploy-manager-admin/src/main/java/no/difi/deploymanager/artifact/ApiController.java
@@ -1,11 +1,11 @@
 package no.difi.deploymanager.artifact;
 
-import no.difi.deploymanager.domain.*;
 import no.difi.deploymanager.domain.Artifact;
+import no.difi.deploymanager.domain.ArtifactList;
+import no.difi.deploymanager.domain.ArtifactType;
 import no.difi.deploymanager.service.ArtifactService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.eclipse.jetty.server.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
