+++ deploy-manager-admin/src/main/java/no/difi/deploymanager/artifact/Beans.java
