+++ deploy-manager-admin/src/main/java/no/difi/deploymanager/artifact/SecurityConfig.java
