+++ deploy-manager-admin/src/main/java/no/difi/deploymanager/domain/ArtifactType.java
