+++ deploy-manager-admin/src/main/java/no/difi/deploymanager/service/ArtifactService.java
