+++ deploy-manager-admin/src/test/java/no/difi/deploymanager/ObjectMotherArtifact.java
