+++ deploy-manager-admin/src/test/java/no/difi/deploymanager/service/ArtifactServiceTest.java
