+++ pom.xml
@@ -4,12 +4,14 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>difi-deploy-manager</groupId>
-  <artifactId>no.difi.deploymanager</artifactId>
+  <groupId>no.difi.deploymanager</groupId>
+  <artifactId>no-difi-deploymanager</artifactId>
   <packaging>pom</packaging>
   <version>0.9.1-SNAPSHOT</version>
   <modules>
     <module>deploy-manager</module>
+    <module>deploy-manager-admin</module>
+    <module>deploy-manager-health-check</module>
   </modules>
 
   <parent>
